Various tweaks and create prodcuers taxonomy
Fixed a few issues I found with syncing and addded the current since parameter to the settings screen for easier debugging. This also creates a producers taxonomy and the default dipslay links to the producer. This nessitated processing the tags so that commas were replaced with dashes since wordpress splits tags on commas