Update the publish date when syncing existing show posts from cablecast
This uses the Cablecast Show's publish date to updte the wordpress posts publish date for show content types. The thoughts behind the original design are documented in #17, but given we have requests for change, and no one strongly advocating for keeping existing behanvior, lets get it working for those that want it and add more config options later if needed.